pyproject.toml
[project]
name = "cbc2026-team3-central"
version = "0.1.0"
description = "CBC2026 Team3 ロボット中央制御プログラム"
requires-python = ">=3.12"
dependencies = [
    "python-can>=4.4",
    "aiohttp>=3.10",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-p no:launch_testing_ros_pytest_entrypoint -p no:launch_testing -p no:launch_ros"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
